+++ Modules/Authentication/src/User.rs
@@ -1,7 +1,15 @@
 use miniserde::{Deserialize, Serialize};
+use File_system::{Flags_type, Mode_type, Open_type, Path_owned_type, Path_type};
 use Users::{
     Group_identifier_inner_type, Group_identifier_type, User_identifier_inner_type,
     User_identifier_type,
+};
+use Virtual_file_system::{File_type, Virtual_file_system_type};
+
+use crate::{
+    Error_type,
+    Hash::{Generate_salt, Hash_password},
+    Result_type, Users_folder_path,
 };
 
 #[derive(Clone, Debug, Deserialize, Serialize)]
@@ -65,4 +73,195 @@ impl User_type {
     pub fn Set_name(&mut self, Name: String) {
         self.Name = Name;
     }
+}
+
+pub fn Get_user_file_path(User_name: &str) -> Result_type<Path_owned_type> {
+    Path_type::New(Users_folder_path)
+        .to_owned()
+        .Append(User_name)
+        .ok_or(Error_type::Failed_to_get_user_file_path)
+}
+
+pub fn Authenticate_user<'a>(
+    Virtual_file_system: &'a Virtual_file_system_type<'a>,
+    User_name: &str,
+    Password: &str,
+) -> Result_type<User_identifier_type> {
+    let Path = Get_user_file_path(User_name)?;
+
+    let User_file = File_type::Open(Virtual_file_system, Path, Mode_type::Read_only.into())
+        .map_err(Error_type::Failed_to_open_user_file)?;
+
+    let mut Buffer = Vec::new();
+
+    User_file
+        .Read_to_end(&mut Buffer)
+        .map_err(Error_type::Failed_to_read_user_file)?;
+
+    let User: User_type = miniserde::json::from_str(core::str::from_utf8(&Buffer).unwrap())
+        .map_err(Error_type::Failed_to_parse_user_file)?;
+
+    if Hash_password(Password, User.Get_salt()) == User.Get_hash() {
+        Ok(User.Get_identifier())
+    } else {
+        Err(Error_type::Invalid_password)
+    }
+}
+
+pub fn Create_user<'a>(
+    Virtual_file_system: &'a Virtual_file_system_type<'a>,
+    User_name: &str,
+    Password: &str,
+    Primary_group: Group_identifier_type,
+    User_identifier: Option<User_identifier_type>,
+) -> Result_type<User_identifier_type> {
+    let Users_manager = Users::Get_instance();
+
+    // - New user identifier if not provided.
+    let User_identifier = if let Some(User_identifier) = User_identifier {
+        User_identifier
+    } else {
+        Users_manager
+            .Get_new_user_identifier()
+            .map_err(Error_type::Failed_to_get_new_user_identifier)?
+    };
+
+    // - Add it to the users manager.
+    Users_manager
+        .Add_user(User_identifier, User_name, Primary_group)
+        .map_err(Error_type::Failed_to_create_user)?;
+
+    // - Hash password.
+    let Salt = Generate_salt()?;
+
+    let Hash = Hash_password(Password, &Salt);
+
+    // - Write user file.
+    let User = User_type::New(
+        User_identifier.Into_inner(),
+        User_name.to_string(),
+        Primary_group.Into_inner(),
+        Hash,
+        Salt,
+    );
+
+    let User_file_path = Path_type::New(Users_folder_path)
+        .to_owned()
+        .Append(User_name)
+        .ok_or(Error_type::Failed_to_get_user_file_path)?;
+
+    let User_file = File_type::Open(
+        Virtual_file_system,
+        User_file_path,
+        Flags_type::New(Mode_type::Write_only, Some(Open_type::Create_only), None),
+    )
+    .map_err(Error_type::Failed_to_open_user_file)?;
+
+    let User_json = miniserde::json::to_string(&User);
+
+    User_file
+        .Write(User_json.as_bytes())
+        .map_err(Error_type::Failed_to_write_user_file)?;
+
+    Ok(User_identifier)
+}
+
+pub fn Change_user_password<'a>(
+    Virtual_file_system: &'a Virtual_file_system_type<'a>,
+    User_name: &str,
+    New_password: &str,
+) -> Result_type<()> {
+    let Salt = Generate_salt()?;
+
+    let Hash = Hash_password(New_password, &Salt);
+
+    let User_file_path = Path_type::New(Users_folder_path)
+        .to_owned()
+        .Append(User_name)
+        .ok_or(Error_type::Failed_to_get_user_file_path)?;
+
+    let User_file = File_type::Open(
+        Virtual_file_system,
+        User_file_path,
+        Flags_type::New(Mode_type::Read_write, Some(Open_type::Truncate), None),
+    )
+    .map_err(Error_type::Failed_to_open_user_file)?;
+
+    let mut Buffer = Vec::new();
+
+    User_file
+        .Read_to_end(&mut Buffer)
+        .map_err(Error_type::Failed_to_read_user_file)?;
+
+    let mut User: User_type = miniserde::json::from_str(core::str::from_utf8(&Buffer).unwrap())
+        .map_err(Error_type::Failed_to_parse_user_file)?;
+
+    User.Set_hash(Hash);
+    User.Set_salt(Salt);
+
+    let User_json = miniserde::json::to_string(&User);
+
+    User_file
+        .Write(User_json.as_bytes())
+        .map_err(Error_type::Failed_to_write_user_file)?;
+
+    Ok(())
+}
+
+pub fn Change_user_name<'a>(
+    Virtual_file_system: &'a Virtual_file_system_type<'a>,
+    Current_name: &str,
+    New_name: &str,
+) -> Result_type<()> {
+    let File_path = Get_user_file_path(Current_name)?;
+
+    let User_file = File_type::Open(
+        Virtual_file_system,
+        File_path,
+        Flags_type::New(Mode_type::Read_write, Some(Open_type::Truncate), None),
+    )
+    .map_err(Error_type::Failed_to_open_user_file)?;
+
+    let mut Buffer = Vec::new();
+
+    User_file
+        .Read_to_end(&mut Buffer)
+        .map_err(Error_type::Failed_to_read_user_file)?;
+
+    let mut User: User_type = miniserde::json::from_str(core::str::from_utf8(&Buffer).unwrap())
+        .map_err(Error_type::Failed_to_parse_user_file)?;
+
+    User.Set_name(New_name.to_string());
+
+    let User_json = miniserde::json::to_string(&User);
+
+    User_file
+        .Write(User_json.as_bytes())
+        .map_err(Error_type::Failed_to_write_user_file)?;
+
+    Ok(())
+}
+
+pub fn Read_user_file<'a>(
+    Virtual_file_system: &'a Virtual_file_system_type<'a>,
+    Buffer: &mut Vec<u8>,
+    File: &str,
+) -> Result_type<User_type> {
+    let User_file_path = Get_user_file_path(File)?;
+
+    let User_file = File_type::Open(
+        Virtual_file_system,
+        User_file_path,
+        Mode_type::Read_only.into(),
+    )
+    .map_err(Error_type::Failed_to_read_users_directory)?;
+
+    Buffer.clear();
+
+    User_file
+        .Read_to_end(Buffer)
+        .map_err(Error_type::Failed_to_read_user_file)?;
+
+    miniserde::json::from_str(core::str::from_utf8(Buffer).unwrap())
+        .map_err(Error_type::Failed_to_parse_user_file)
 }
